\docType{data}
\name{chinook_run_settings}
\alias{chinook_run_settings}
\title{Parameter settings for long runs of the chinook data.}
\format{This is a data frame with 6 rows and 5 columns.  Each row contains a set of
parameters to be passed to do_chinook().  It's just there to make it easy to
set up some runs in parallel.}
\usage{
data(chinook_run_settings)
}
\description{
Parameter settings for long runs of the chinook data.
}
\keyword{datasets}

