\name{do_chinook_stock}
\alias{do_chinook_stock}
\title{Do the chinook stock analysis in the current working directory}
\usage{
do_chinook_stock(nsim, B, DoGSI = FALSE)
}
\arguments{
  \item{nsim}{number of simulations}

  \item{B}{number of bootstrap replicates}

  \item{DoGSI}{True means do the GSI simulation}
}
\description{
Do the chinook stock analysis in the current working directory
}

