\name{do_steelhead_stock}
\alias{do_steelhead_stock}
\title{Do the steelhead stock analysis in the current working directory}
\usage{
do_steelhead_stock(nsim, B, DoGSI = FALSE)
}
\arguments{
  \item{nsim}{number of simulations}

  \item{B}{number of bootstrap replicates}

  \item{DoGSI}{True means do the GSI simulation}
}
\description{
Do the steelhead stock analysis in the current working directory
}

