% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower-granite-gsi-funcs.R
\name{extract_multi_fix_sim_to_df}
\alias{extract_multi_fix_sim_to_df}
\title{read in gsi_sim ouput and extract the multi-fix-mix lines into a data frame}
\usage{
extract_multi_fix_sim_to_df(gsi.big.out, ru.list, Originnames)
}
\arguments{
\item{gsi.big.out}{A character vector of gsi_sim's standard output.}
}
\description{
takes a character vector of gsi_sim output from readLines, extracts the multi-fix-mix 
output lines and gets the header that
it needs to make a huge data frame of them all, then, using the popnames
and the ru.list, finds the MLE reporting group for each individual.
}

