% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_pop_for_rg}
\alias{get_pop_for_rg}
\title{return a population drawn at random from those in a reporting unit}
\usage{
get_pop_for_rg(ru.list, sampled_rgs)
}
\arguments{
\item{ru.list}{list whose names are reporting units and components are vectors of population names}

\item{sampled_rgs}{a vector of reporting units (denoting the stock of each fish for example)}
}
\description{
pass this a vector of reporting group names and a list of the
populations in each of those, and this creates a vector of populations
sampled uniformly from within each rg
}

