% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower-granite-gsi-funcs.R
\name{gsi_ize_the_Sim.List}
\alias{gsi_ize_the_Sim.List}
\title{modify the Sim.List so that Prop are gsi assignments}
\usage{
gsi_ize_the_Sim.List(Sim.List, ru.list, GSISIM, BLFILE, Originnames, BL.pops)
}
\description{
modify the Sim.List so that Prop are gsi assignments
}

