% Generated by roxygen2 (4.0.0): do not edit by hand

\name{gsisimRepUnits2List}
\alias{gsisimRepUnits2List}
\title{reads a gsi_sim reporting units file and returns and list of reporting units}
\usage{

  gsisimRepUnits2List(ru.file = "/Users/eriq/Documents/xp_dev_svn_checkouts/gsi_sim/snpset/2010_SNPset_GSI_TOOLS/Baseline/snpset_ReportingUnits.txt")
}
\description{
They are ordered as they are ordered in the rep-units file.  Names of the components
are the names of the reporting units and the constituent populations are the
contained character vector, in the order they appear in the reporting units file
}

