% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{lg_test_data}
\alias{lg_test_data}
\title{A variety of data that is useful for testing functions in the lowergranite package}
\format{This is a list with the following components:
\describe{
 \item{RU.list}{An example reporting unit list.}
 \item{Originnames}{An example of an Originnames character vector.}
 \item{big.gsi.out}{Example gsi_sim standard output (with --multi-fix-mix output) in a character vector.}
 \item{multi_fix_df_compare}{Data frame: correct output of \code{\link{extract_multi_fix_sim_to_df}} when run with the
 test data above. Allows comparison in tests. This data frame was created with
 \code{expect_that(extract_multi_fix_sim_to_df(lg_test_data$gsi.big.out, lg_test_data$RU.list, lg_test_data$Originnames)}}
 \item{known_stock_result15}{output from run_boot_gsi_analysis.R that should look like some test results.}
 \item{gsi_stock_result15}{output from run_boot_gsi_analysis.R that should look like some test results.}
 \item{SH11.Stock.Age.With.Periods}{A data frame of the 2011 steelhead proportions broken out by stock and brood year with names
 formatted the way we need them for theses analyses.  i.e., like GRROND..BY07 GRROND..BY08 LSNAKE..BY04, etc}
 \item{Sim.List.Vanilla}{This holds the result of the Sim.List variable when running \code{\link{run_boot_gsi_analysis}}
 on the simple SH11 data set that is listed in the example:  i.e. the standard
 \code{known_stock_result1 <- run_boot_gsi_analysis(stock_group_start_col = 14, nsim = 10, B = 50, DO_GSI_ON_PROP = F)}}
 \item{Sim.List.SH.Stock.Age}{This holds the Sim.List with nsim=2 running run_boot_gsi_analysis with W=SH11.Stock.Age.With.Periods.
 This is primarily used to be able to feed it into \code{\link{gsi_ize_the_Sim.List}} to make sure that function is dealing with
 the two-period group names like GRROND..BY07.}
}}
\usage{
data(lg_test_data)
}
\description{
This is a list that holds a variety of quantities that are useful to have
to carry out tests of the functions in the package.  Mostly they were
created from the steelhead data.
}
\keyword{datasets}

