% Generated by roxygen2 (4.0.0): do not edit by hand
\name{make_super_informo_data}
\alias{make_super_informo_data}
\title{create fake "super-informative-data" for a set of baseline pops}
\usage{
make_super_informo_data(BLFILE, outfile = "super_informo_baseline.txt",
  nperpop = 100)
}
\arguments{
  \item{BLFILE}{a baseline file in gsi_sim format that has
  all the population names that you want in the order that
  you want them}

  \item{outfile}{path to where you want the output written}
}
\description{
Output is written as a gsi_sim file.
}

