% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi-sim-r-stuff.R
\name{quoteProtect}
\alias{quoteProtect}
\title{given a file path x (or vector of them) return the path with quotes around it}
\usage{
quoteProtect(x)
}
\arguments{
\item{x}{a file path (or a vector of paths)}
}
\description{
The system path on Windows has spaces in it which causes problems with 
\code{\link{pipe}} and other such system calls.  This just puts those
paths inside quotation marks
}

