% Generated by roxygen2 (4.0.0): do not edit by hand
\name{run_chinook_script}
\alias{run_chinook_script}
\title{Run the chinook script}
\usage{
run_chinook_script(WORK.DIR = getwd(),
  STOCK.DATA.XLSX = "S:/Eagle Fish Genetics Lab/Ackerman/LGR/Kirk/ms/stock_sims/inputs/CH11SIMPOP.xlsx",
  DO_GSI = TRUE,
  BLFILE = "C:/GenSoftware/gsi_sim_pc/analysis/chnk_base_v3_180.txt",
  RUFILE = "C:/GenSoftware/gsi_sim_pc/analysis/chnk_base_v3_repunits.txt",
  GSISIM = "C:/GenSoftware/gsi_sim_pc/gsi_sim.exe",
  GSISIM.FOLDER = "C:/GenSoftware/gsi_sim_pc/analysis/")
}
\arguments{
  \item{WORK.DIR}{The working directory to do this in.
  Default = current working directory}

  \item{STOCK.DATA.XLSX}{the file that has the stock data
  in it used to drive the simulations}

  \item{DO_GSI}{if set to TRUE then it uses GSI
  assignments. Otherwise it uses the true origins}
}
\description{
This is kirk's script for chinook that eric wrapped up in a function
}

