% Generated by roxygen2 (4.0.0): do not edit by hand
\name{run_steelhead_script}
\alias{run_steelhead_script}
\title{Run the steelhead script}
\usage{
run_steelhead_script(DAT.DIR = system.file("data_files", package =
  "lowergranite", mustWork = T), WORK.DIR = getwd(),
  STOCK.DATA.XLSX = file.path(DAT.DIR, "SH11SIMPOPstock.xlsx"),
  DO_GSI_ON_PTABLE = TRUE, DO_GSI_ON_PROP = FALSE,
  GSISIM = "C:/GenSoftware/gsi_sim_pc/gsi_sim.exe",
  BLFILE = file.path(DAT.DIR, "sthd_base_v3_187.txt"),
  RUFILE = file.path(DAT.DIR, "sthd_base_v3_rg.txt"),
  GSISIM.FOLDER = getwd(), alph = 0.1, B = 500, nsim = 500)
}
\arguments{
  \item{DAT.DIR}{The directory where all the data files
  are.  Defaults to the dictory "data_files" in the
  installed package}

  \item{WORK.DIR}{The working directory to do this in.
  Default = current working directory}

  \item{STOCK.DATA.XLSX}{the file that has the stock data
  in it used to drive the simulations}

  \item{DO_GSI_ON_PTABLE}{if set to TRUE then the script
  simulates gsi_sim assignments from the tabulated counts
  of fish from different stocks in each stratum. If this
  and \code{DO_GSI_ON_PROP} are both FALSE, it uses the
  true origins.}

  \item{DO_GSI_ON_PROP}{if set to TRUE then gsi_sim is used
  to create assignments that replace the assignments in the
  variable Prop.  This is required if doing stock-sex or
  stock-age combinations.}

  \item{alph}{One minus the size of the desired confidence
  interval to be calculated}

  \item{B}{number of bootstrap replicates}

  \item{nsim}{number of simulation replicates to do}
}
\description{
This is kirk's script that eric wrapped up in a function
}

