\docType{data}
\name{steelhead_run_settings}
\alias{steelhead_run_settings}
\title{Parameter settings for long runs of the steelhead data.}
\format{This is a data frame with 6 rows and 5 columns.  Each row contains a set of
parameters to be passed to do_steelhead().  It's just there to make it easy to
set up some runs in parallel.}
\usage{
data(steelhead_run_settings)
}
\description{
Parameter settings for long runs of the steelhead data.
}
\keyword{datasets}

