% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gsi_ize_the_Sim.List}
\alias{gsi_ize_the_Sim.List}
\title{modify the Sim.List so that Prop are gsi assignments}
\usage{
gsi_ize_the_Sim.List(Sim.List, ru.list, GSISIM, BLFILE, Originnames, BL.pops)
}
\description{
modify the Sim.List so that Prop are gsi assignments
}

